refactor: move settings.ts to constants/
