+++ src/features/machines/withdrawUIMachine.ts
@@ -7,7 +7,7 @@ import {
   setup,
   spawnChild,
 } from "xstate"
-import { settings } from "../../config/settings"
+import { settings } from "../../constants/settings"
 import { logger } from "../../logger"
 import type { QuoteResult } from "../../services/quoteService"
 import type { BaseTokenInfo, UnifiedTokenInfo } from "../../types/base"
