+++ src/features/otcDesk/components/OtcMakerTrades.tsx
@@ -22,7 +22,7 @@ import * as v from "valibot"
 import { type ActorRefFrom, createActor, toPromise } from "xstate"
 import { AssetComboIcon } from "../../../components/Asset/AssetComboIcon"
 import { Copy } from "../../../components/IntentCard/CopyButton"
-import { settings } from "../../../config/settings"
+import { settings } from "../../../constants/settings"
 import type { DefuseUserId, SignerCredentials } from "../../../core/formatters"
 import { getDepositedBalances } from "../../../services/defuseBalanceService"
 import type { BaseTokenInfo, UnifiedTokenInfo } from "../../../types/base"
