+++ src/features/otcDesk/components/OtcTakerWidget.tsx
@@ -5,7 +5,7 @@ import type { CodeResult } from "near-api-js/lib/providers/provider"
 import { type ReactNode, useMemo, useState } from "react"
 import * as v from "valibot"
 import { WidgetRoot } from "../../../components/WidgetRoot"
-import { settings } from "../../../config/settings"
+import { settings } from "../../../constants/settings"
 import type { DefuseUserId, SignerCredentials } from "../../../core/formatters"
 import { logger } from "../../../logger"
 import { SwapWidgetProvider } from "../../../providers/SwapWidgetProvider"
